+++ src/components/nav/index.ts
@@ -1,0 +1,2 @@
+export { default as Navbar } from "./Navbar";
+export { default as FloatingCartButton } from "./FloatingCartButton";
